feat: add stories to the Home page
Created storyBubble and addStoryButton components.

refactor: from now on the people's datas are in an array and listed out into a FlatList so we can scroll trough them.
Also implemented inside the chatBubble component that if the status is offline the status icon does not show up, and if the lastMsg is too long it's trunk down the msg and adds '...' tail to it.